Started Board and stone.
Coordinate is more of a static class to get unique coordinate integers.
This assures that x 1 + y 11 isnt the same as x 11 + y 1.
May we can create a class on top of HashMap that overrides the get and
takes a x, y instead of integer. (maybe to much, this works too)