Refactors BasicTUI and drawBoard to be implemented in the Log class
As a server and client both have a TUI it is important both have access
to the drawBoard function. Furthermore, the TUI should not be too busy
figuring out what to draw, that should be handled by the
(Server/Client)Log. This means that if we were to replace the TUI with
a GUI it would be very easy because of the observer pattern. TUI should
not contain too much app specific logic.